Multiplatform support

+++ include/file_utils.sh
@@ -23,7 +23,7 @@ file_replace() {
 	[ -z "$3" ] && exit # text to replace regex to
 	#echo "s/$2/$3/g$4"
 	sed -i.bac -E "s/$2/$3/g$4" $1
-	rm $1.bac
+	[ -f "$1.bac" ] && rm $1.bac
 	# Use python due to platform independence
 	# use relative paths due to platform independence
 	# fpath=$(realpath --relative-to="${PWD}" "$1")
